Reorganize code into a package
Create the `scribe` package that contains the most basic tools for writing text documents
